+++ wix/atlascli.json
@@ -4,7 +4,7 @@
   "license": "LICENSE",
   "upgrade-code": "a995cda5-9731-4158-910c-c8d15371dbda",
   "files": {
-    "guid": "c7b0df57-c3c1-4591-becd-f0f73a912eb8",
+    "guid": "ad7bb4a1-9026-4578-8dff-400d66a02a51",
     "items": [
       "bin/atlas.exe"
     ]
